consolidate pytes
