--- python/tests/generic_alias_2.py
@@ -1,22 +1,0 @@
-from typing import Literal, TypedDict, Generic, TypeVar
-from typing_extensions import TypeAliasType
-
-T = TypeVar("T", covariant=True)
-
-
-class First(Generic[T], TypedDict):
-    kind: Literal["first"]
-    first_attr: T
-
-
-class Second(Generic[T], TypedDict):
-    kind: Literal["second"]
-    second_attr: T
-
-
-FirstOrSecond = TypeAliasType("FirstOrSecond", First[T] | Second[T], type_params=(T,))
-
-
-class Nested(TypedDict):
-    item: FirstOrSecond[str]
-
