+++ python/tests/test_generic_alias_2.py
@@ -1,6 +1,26 @@
-from typing_extensions import Any
+from typing_extensions import TypeAliasType, Any
+from typing import Literal, TypedDict, Generic, TypeVar
 from typechat import python_type_to_typescript_schema
-from tests.generic_alias_2 import Nested
+
+T = TypeVar("T", covariant=True)
+
+class First(Generic[T], TypedDict):
+    kind: Literal["first"]
+    first_attr: T
+
+
+class Second(Generic[T], TypedDict):
+    kind: Literal["second"]
+    second_attr: T
+
+
+FirstOrSecond = TypeAliasType("FirstOrSecond", First[T] | Second[T], type_params=(T,))
+
+
+class Nested(TypedDict):
+    item: FirstOrSecond[str]
+
+
 
 def test_generic_alias1(snapshot : Any):
     assert(python_type_to_typescript_schema(Nested) == snapshot)
